requirements.txt
albucore==0.0.23
albumentations==2.0.0
annotated-types==0.7.0
apt-clone==0.2.1
apturl==0.5.2
beautifulsoup4==4.10.0
blinker==1.4
Brlapi==0.8.3
certifi==2020.6.20
chardet==4.0.0
click==8.0.3
colorama==0.4.4
command-not-found==0.3
configobj==5.0.6
contourpy==1.3.1
cryptography==3.4.8
cupshelpers==1.0
cycler==0.12.1
dbus-python==1.2.18
defer==1.0.6
distro==1.7.0
eyeD3==0.8.10
filelock==3.6.0
fonttools==4.55.3
fsspec==2024.12.0
grpcio==1.30.2
gyp==0.1
html5lib==1.1
httplib2==0.20.2
idna==3.3
ifaddr==0.1.7
IMDbPY==2021.4.18
ipython_genutils==0.2.0
jeepney==0.7.1
Jinja2==3.1.5
jupyter-core==4.9.1
kiwisolver==1.4.8
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
louis==3.20.0
lxml==4.8.0
macaroonbakery==1.3.1
Mako==1.1.3
MarkupSafe==2.0.1
matplotlib==3.10.0
meld==3.20.4
more-itertools==8.10.0
mpmath==1.3.0
nemo-emblems==5.8.0
netaddr==0.8.0
netifaces==0.11.0
networkx==3.4.2
numpy==2.1.2
nvidia-cublas-cu12==12.4.5.8
nvidia-cuda-cupti-cu12==12.4.127
nvidia-cuda-nvrtc-cu12==12.4.127
nvidia-cuda-runtime-cu12==12.4.127
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.2.1.3
nvidia-curand-cu12==10.3.5.147
nvidia-cusolver-cu12==11.6.1.9
nvidia-cusparse-cu12==12.3.1.170
nvidia-nccl-cu12==2.21.5
nvidia-nvjitlink-cu12==12.4.127
nvidia-nvtx-cu12==12.4.127
oauthlib==3.2.0
onboard==1.4.1
opencv-python==4.10.0.84
opencv-python-headless==4.10.0.84
packaging==21.3
PAM==0.4.2
pexpect==4.8.0
Pillow==9.0.1
protobuf==3.12.4
psutil==5.9.0
ptyprocess==0.7.0
pycairo==1.20.1
pycocotools==2.0.8
pycups==2.0.1
pycurl==7.44.1
pydantic==2.10.5
pydantic_core==2.27.2
pyelftools==0.27
Pygments==2.11.2
PyGObject==3.42.1
PyICU==2.8.1
pyinotify==0.9.6
PyJWT==2.3.0
pymacaroons==0.13.0
PyNaCl==1.5.0
pyparsing==2.4.7
pyparted==3.11.7
pyRFC3339==1.1
python-apt==2.4.0+ubuntu4
python-dateutil==2.9.0.post0
python-debian==0.1.43+ubuntu1.1
python-gnupg==0.4.8
python-magic==0.4.24
python-xlib==0.29
pytz==2022.1
PyYAML==5.4.1
reportlab==3.6.8
requests==2.25.1
requests-file==1.5.1
scipy==1.15.1
SecretStorage==3.3.1
setproctitle==1.2.2
simsimd==6.2.1
six==1.16.0
soupsieve==2.3.1
stringzilla==3.11.3
sympy==1.13.1
systemd-python==234
tinycss2==1.1.1
tldextract==3.1.2
torch==2.5.1
torchvision==0.20.1
traitlets==5.1.1
triton==3.1.0
typing_extensions==4.12.2
ubuntu-drivers-common==0.0.0
ufw==0.36.1
Unidecode==1.3.3
urllib3==1.26.5
wadllib==1.3.6
webencodings==0.5.1
wxPython==4.2.0
xkit==0.0.0
xlrd==1.2.0
youtube-dl==2021.12.17
